add function and maybe there is a problem

+++ NeuralNet/Layer.cpp
@@ -16,7 +16,7 @@ void Layer::init(int numNeurons, int nextNumNeuron, int indexLayer)
 	neurons.resize(numNeurons);
 	index = indexLayer;
 
-	for (size_t i = 0; i < numNeurons; i++)
+	for (int i = 0; i < numNeurons; i++)
 	{
 		neurons[i].index = i;
 		neurons[i].weights.resize(nextNumNeuron);
